+++ test/exceeds_expectations/test/expect_not/to_throw.d
@@ -4,7 +4,7 @@ import exceeds_expectations;
 import exceeds_expectations.test;
 
 
-@("toThrow, since it throws the wrong thing")
+@("succeed when it throws the wrong thing")
 unittest
 {
     class CustomException : Exception
@@ -18,13 +18,13 @@ unittest
     expect({ throw new Exception("test"); }).not.toThrow!CustomException;
 }
 
-@("toThrow, since it doesn't throw anything")
+@("succeed when it doesn't throw anything")
 unittest
 {
     expect({ return; }).not.toThrow!Exception;
 }
 
-@("toThrow, but throws the exact thing")
+@("fail when it throws the exact thing")
 unittest
 {
     shouldFail(
@@ -32,7 +32,7 @@ unittest
     );
 }
 
-@("toThrow, but throws a sub-type")
+@("fail when it throws a sub-type")
 unittest
 {
     class CustomException : Exception
@@ -45,5 +45,58 @@ unittest
 
     shouldFail(
         expect({ throw new CustomException("Test"); }).not.toThrow
+    );
+}
+
+@("succeed if it throws the right type but the message does not match")
+unittest
+{
+    class UnexpectedValueException : Exception
+    {
+        this(string msg, string file = __FILE__, size_t line = __LINE__, Throwable nextInChain = null)
+        pure nothrow @safe
+        {
+            super("Unexpected value: " ~ msg, file, line, nextInChain);
+        }
+    }
+
+    expect({
+        throw new UnexpectedValueException("spanish inquisition");
+    }).not.toThrow!UnexpectedValueException("something else");
+}
+
+@("succeed if the message matches but the type is wrong")
+unittest
+{
+    class UnexpectedValueException : Exception
+    {
+        this(string msg, string file = __FILE__, size_t line = __LINE__, Throwable nextInChain = null)
+        pure nothrow @safe
+        {
+            super("Unexpected value: " ~ msg, file, line, nextInChain);
+        }
+    }
+
+    expect({
+        throw new UnexpectedValueException("spanish inquisition");
+    }).not.toThrow!Error("spanish inquisition");
+}
+
+@("fail if the type and message both match")
+unittest
+{
+    class UnexpectedValueException : Exception
+    {
+        this(string msg, string file = __FILE__, size_t line = __LINE__, Throwable nextInChain = null)
+        pure nothrow @safe
+        {
+            super("Unexpected value: " ~ msg, file, line, nextInChain);
+        }
+    }
+
+    shouldFail(
+        expect({
+            throw new UnexpectedValueException("spanish inquisition");
+        }).toThrow!UnexpectedValueException("spanish inquisition")
     );
 }